task4/requirements.txt
scipy==1.7.0
pandas==1.2.5
tqdm==4.59.0
fire==0.4.0
numpy==1.20.3
torch==1.8.1
torch_audiomentations==0.8.0
pytorch-ignite==0.4.5
loguru==0.5.3
h5py==3.2.1
einops==0.3.0
torchaudio==0.8.1
psds_eval==0.4.0
PyYAML==5.4.1
scikit_learn==0.24.2
sed_eval==0.2.1
